Extract I/O to zrad.io and add Image factories
Move NIfTI/DICOM I/O and PET normalization logic out of zrad/image.py into a new zrad.io package (dicom, nifti, pet_suv). Replace instance read_* methods with Image class factories (from_nifti, from_nifti_mask, from_dicom, from_dicom_mask) and an internal _from_sitk_image. Update GUI modules and tests to import I/O helpers from zrad.io.dicom and to use the new Image.from_* factories. Update docs example to demonstrate Image.from_nifti/from_nifti_mask usage.
