+++ zrad/gui/prep_tab.py
@@ -21,7 +21,7 @@ from .toolbox_gui import (
     ProcessingWorker,
 )
 from ..exceptions import InvalidInputParametersError, DataStructureError
-from ..image import get_all_structure_names, get_dicom_files
+from ..io.dicom import get_all_structure_names, get_dicom_files
 from ..preprocessing import Resampler
 from ..toolbox_logic import get_logger, close_all_loggers, joblib_progress
