+++ zrad/io/__init__.py
@@ -1,0 +1,18 @@
+"""Image input/output helpers."""
+
+from .dicom import (
+    get_all_structure_names,
+    get_dicom_files,
+    read_dicom_image,
+    read_dicom_mask,
+)
+from .nifti import read_nifti_image, read_nifti_mask
+
+__all__ = [
+    "get_all_structure_names",
+    "get_dicom_files",
+    "read_dicom_image",
+    "read_dicom_mask",
+    "read_nifti_image",
+    "read_nifti_mask",
+]
